# openapi.yaml
openapi: 3.0.3
info:
  title: QR Reward Platform API
  version: '1.0'
  description: |
    Promoters fund QR campaigns; end users scan and sign up with a publisher; coins move
    automatically, exactly once. See `SYSTEM_FLOW.md` for the end-to-end narrative behind
    these endpoints and the invariants each one enforces.

    ### Auth schemes
    - **sessionAuth** — `Authorization: Bearer <jwt>` from `/v1/auth/login` or `/v1/auth/signup`.
      12h expiry. Used by promoter/publisher portal routes and the admin console.
    - **apiKeyAuth** — `Authorization: Bearer pk_<hex>` issued once at publisher signup (or via
      rotate). Used only by the publisher's own server, never a browser.
    - Unauthenticated routes: `/healthz`, `GET /r/:code`, `GET /v1/qr-codes/:id/image`,
      `POST /v1/auth/signup`, `POST /v1/auth/login`.
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: '{baseUrl}'
    description: Deployed
    variables:
      baseUrl:
        default: https://api.example.com
security: []
...
paths:
...
  /v1/publishers:
    get:
      tags: [Portal]
      summary: List publishers (for choosing a partnership target)
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties: { id: { type: string, format: uuid }, name: { type: string } }
  /v1/partnerships:
    post:
      tags: [Portal]
      summary: Request a partnership with a publisher
      description: Promoter only. Sets the commercial terms for every campaign under it.
      security: [{ sessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [publisher_org_id]
              properties:
                publisher_org_id: { type: string, format: uuid }
                coin_rate: { type: integer, minimum: 1, maximum: 100000, default: 50 }
                guest_rate: { type: integer, minimum: 0, description: 'Must be <= coin_rate', default: 10 }
                grace_days: { type: integer, minimum: 0, maximum: 365, default: 7 }
      responses:
        '201':
          description: Partnership created with status `pending`
          content: { application/json: { schema: { $ref: '#/components/schemas/Partnership' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [Portal]
      summary: List partnerships involving my org
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Partnership' } }
  /v1/partnerships/{id}/accept:
    post:
      tags: [Portal]
      summary: Accept a pending partnership request
      description: Publisher only, and only the publisher named on that partnership.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Partnership now `active`
          content: { application/json: { schema: { $ref: '#/components/schemas/Partnership' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/campaigns:
    post:
      tags: [Portal]
      summary: Create a campaign under an active partnership
      description: Promoter only.
      security: [{ sessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partnership_id, name]
              properties:
                partnership_id: { type: string, format: uuid }
                name: { type: string }
      responses:
        '201':
          description: Campaign created with status `active` and zero budget
          content: { application/json: { schema: { $ref: '#/components/schemas/Campaign' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      tags: [Portal]
      summary: List campaigns visible to my org
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Campaign' } }
  /v1/campaigns/{id}/fund:
    post:
      tags: [Portal]
      summary: Fund a campaign's budget
      description: |
        Promoter only, owner of the campaign. Demo funding credits the ledger directly
        (`external:funding` → `campaign:{id}`); production would put a PSP checkout in front.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [coins]
              properties: { coins: { type: integer, minimum: 1, maximum: 10000000 } }
      responses:
        '201':
          description: New budget
          content:
            application/json:
              schema: { type: object, properties: { budget: { type: integer } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/campaigns/{id}:
    patch:
      tags: [Portal]
      summary: Pause, resume, or end a campaign
      description: Promoter only, owner of the campaign.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { type: string, enum: [active, paused, ended] } }
      responses:
        '200':
          description: Updated campaign
          content: { application/json: { schema: { $ref: '#/components/schemas/Campaign' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/campaigns/{id}/stats:
    get:
      tags: [Portal]
      summary: Campaign performance — scans, redemptions, coins granted, budget left
      description: Either party on the partnership may read.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  scans: { type: integer }
                  redemptions: { type: integer }
                  coins_granted: { type: integer }
                  budget_remaining: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/campaigns/{id}/qr-codes:
    post:
      tags: [Portal]
      summary: Create a QR code for a campaign
      description: Promoter only, owner of the campaign. Time-bound by default (30 days).
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                style: { $ref: '#/components/schemas/QrStyle' }
                expires_in_days:
                  type: integer
                  minimum: 0
                  maximum: 3650
                  default: 30
                  description: '0 = never expires'
                max_uses:
                  type: integer
                  minimum: 1
                  nullable: true
                  description: 'Omit/null = unlimited'
      responses:
        '201':
          description: QR code created
          content: { application/json: { schema: { $ref: '#/components/schemas/QrCode' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      tags: [Portal]
      summary: List QR codes for a campaign
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/QrCode' } }
  /v1/qr-codes/{id}/void:
    post:
      tags: [Portal]
      summary: Void a QR code (lost/stolen print run)
      description: |
        Promoter only, owner of the code. Promoters may only tighten a code (void it), never
        loosen one — extending expiry or use limits is an admin-only override.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Voided
          content: { application/json: { schema: { $ref: '#/components/schemas/QrCode' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/qr-codes/{id}:
    patch:
      tags: [Portal]
      summary: Restyle a QR code
      description: Promoter only, owner of the code. Cosmetic only — does not touch expiry or use limits.
      security: [{ sessionAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [style]
              properties: { style: { $ref: '#/components/schemas/QrStyle' } }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/QrCode' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/api-keys/rotate:
    post:
      tags: [Portal]
      summary: Rotate my Partner API key
      description: Publisher only. Old key is invalidated immediately; new key shown once.
      security: [{ sessionAuth: [] }]
      responses:
        '201':
          description: New key
          content:
            application/json:
              schema: { type: object, properties: { api_key: { type: string } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/orgs/me:
    get:
      tags: [Portal]
      summary: My org profile
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Org' } } }
    patch:
      tags: [Portal]
      summary: Update my landing URL
      description: Publisher setting — where scans redirect to. Validated https-only (http for localhost).
      security: [{ sessionAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { landing_url: { type: string } }
      responses:
        '200':
          description: Updated
          content: { application/json: { schema: { $ref: '#/components/schemas/OrgSummary' } } }
  /v1/redemptions:
    get:
      tags: [Portal]
      summary: My org's recent redemptions (portal view)
      description: Latest 100, either side of the partnership.
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Redemption' } }

  # ---------------- Partner API ----------------
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 12h session token from `/v1/auth/signup` or `/v1/auth/login`
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: 'pk_<hex>'
      description: Publisher's Partner API key — server-to-server only
  responses:
    BadRequest:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Forbidden:
      description: Authenticated, but not allowed to perform this action
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Not found (or not owned by the caller — ownership failures read as 404, not 403)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        statusCode: { type: integer }
        message: { type: string }
        error: { type: string }
    OrgSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type: { type: string, enum: [promoter, publisher, admin] }
    Org:
      allOf:
        - { $ref: '#/components/schemas/OrgSummary' }
        - type: object
          properties:
            email: { type: string, format: email }
            landing_url: { type: string, nullable: true }
            suspended: { type: boolean }
...
    Partnership:
      type: object
      properties:
        id: { type: string, format: uuid }
        promoter_org_id: { type: string, format: uuid }
        publisher_org_id: { type: string, format: uuid }
        promoter_name: { type: string }
        publisher_name: { type: string }
        coin_rate: { type: integer }
        guest_rate: { type: integer }
        grace_days: { type: integer }
        status: { type: string, enum: [pending, active] }
        created_at: { type: string, format: date-time }
    Campaign:
      type: object
      properties:
        id: { type: string, format: uuid }
        partnership_id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [active, paused, ended] }
        coin_rate: { type: integer }
        promoter_name: { type: string }
        publisher_name: { type: string }
        budget: { type: integer }
        created_at: { type: string, format: date-time }
    QrStyle:
      type: object
      description: Colors, size, quiet zone, error correction, optional center logo
      properties:
        fgColor: { type: string }
        bgColor: { type: string }
        size: { type: integer }
        margin: { type: integer }
        errorCorrectionLevel: { type: string, enum: [L, M, Q, H] }
        logo: { type: string, nullable: true, description: Data URL }
    QrCode:
      type: object
      properties:
        id: { type: string, format: uuid }
        campaign_id: { type: string, format: uuid }
        code: { type: string }
        scan_url: { type: string }
        style: { $ref: '#/components/schemas/QrStyle' }
        expires_at: { type: string, format: date-time, nullable: true }
        max_uses: { type: integer, nullable: true }
        uses: { type: integer }
        voided: { type: boolean }
        created_at: { type: string, format: date-time }
    Redemption:
      type: object
      properties:
        id: { type: string, format: uuid }
        campaign_id: { type: string, format: uuid }
        campaign_name: { type: string }
        scan_id: { type: string, format: uuid }
        publisher_user_ref: { type: string }
        coins: { type: integer }
        identified: { type: boolean }
        upgraded_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...